Windows port of antirez/3.2.

+++ tests/instances.tcl
@@ -56,7 +56,11 @@ proc exec_instance {type cfgfile} {
 # Spawn a redis or sentinel instance, depending on 'type'.
 proc spawn_instance {type base_port count {conf {}}} {
     for {set j 0} {$j < $count} {incr j} {
-        set port [find_available_port $base_port]
+        if { $::tcl_platform(platform) == "windows" } {
+            set port ${base_port}
+        } else {
+            set port [find_available_port $base_port]
+        }
         incr base_port
         puts "Starting $type #$j at port $port"
 
@@ -79,7 +83,13 @@ proc spawn_instance {type base_port count {conf {}}} {
         close $cfg
 
         # Finally exec it and remember the pid for later cleanup.
-        set pid [exec_instance $type $cfgfile]
+        if {$type eq "redis"} {
+            set pid [exec ../../../src/redis-server $cfgfile &]             ;# WIN_PORT_FIX
+        } elseif {$type eq "sentinel"} {
+            set pid [exec ../../../src/redis-server $cfgfile --sentinel &]  ;# WIN_PORT_FIX
+        } else {
+            error "Unknown instance type."
+        }
         lappend ::pids $pid
 
         # Check availability
@@ -119,7 +129,13 @@ proc cleanup {} {
     puts "Cleaning up..."
     log_crashes
     foreach pid $::pids {
-        catch {exec kill -9 $pid}
+        catch {
+			if { $::tcl_platform(platform) == "windows" } {
+				kill_proc2 $pid
+			} else {
+				exec kill -9 $pid
+			}
+		}
     }
     foreach dir $::dirs {
         catch {exec rm -rf $dir}
@@ -446,8 +462,11 @@ proc kill_instance {type id} {
     if {$pid == -1} {
         error "You tried to kill $type $id twice."
     }
-
-    exec kill -9 $pid
+    if { $::tcl_platform(platform) == "windows" } {
+        kill_proc2 $pid
+    } else {
+        exec kill -9 $pid
+    }
     set_instance_attrib $type $id pid -1
     set_instance_attrib $type $id link you_tried_to_talk_with_killed_instance
 
@@ -482,7 +501,11 @@ proc restart_instance {type id} {
 
     # Execute the instance with its old setup and append the new pid
     # file for cleanup.
-    set pid [exec_instance $type $cfgfile]
+    if {$type eq "redis"} {
+        set pid [exec ../../../src/redis-server $cfgfile &]               ;# WIN_PORT_FIX
+    } else {
+        set pid [exec ../../../src/redis-server $cfgfile --sentinel &]    ;# WIN_PORT_FIX
+    }
     set_instance_attrib $type $id pid $pid
     lappend ::pids $pid
